refactor: move formatTime into Clockface component
