+++ src/components/clock/Clockface.js
@@ -1,6 +1,18 @@
 import React from 'react';
 import {Digit} from './Digit.js';
-import {formatTime} from '../../lib/clockHelpers';
+
+const formatTime = (dateObject) => {
+  let formattedTime = formatLeadingZero(dateObject.getHours()) + ':';
+  formattedTime += formatLeadingZero(dateObject.getMinutes()) + ':';
+  formattedTime += formatLeadingZero(dateObject.getSeconds());
+
+  return formattedTime;
+
+};
+
+const formatLeadingZero = (num) => {
+  return num < 10 ? '0' + num.toString() : num.toString();
+};
 
 export const Clockface = ({dateObject}) => {
